+++ src/main/java/org/vesalainen/parser/util/CharInput.java
@@ -73,9 +73,9 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
         }
     }
     @Override
-    public int get(int index)
+    public int get(long index)
     {
-        return buffer1.get(index % size);
+        return buffer1.get((int)(index % size));
     }
 
     @Override
@@ -99,13 +99,13 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
     }
 
     @Override
-    protected void set(int index, int value)
+    protected void set(long index, int value)
     {
-        buffer1.put(index % size, (char)value);
+        buffer1.put((int)(index % size), (char)value);
     }
 
     @Override
-    public void write(int start, int length, Writer writer) throws IOException
+    public void write(long start, int length, Writer writer) throws IOException
     {
         if (array != null)
         {
@@ -113,8 +113,8 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
             {
                 throw new IllegalArgumentException("buffer too small");
             }
-            int ps = start % size;
-            int es = (start+length) % size;
+            int ps = (int) (start % size);
+            int es = (int) ((start+length) % size);
             if (ps <= es)
             {
                 writer.write(array, ps, length);
@@ -152,7 +152,7 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
      * @return 
      */
     @Override
-    public String getString(int start, int length)
+    public String getString(long start, int length)
     {
         if (length == 0)
         {
@@ -160,8 +160,8 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
         }
         if (array != null)
         {
-            int ps = start % size;
-            int es = (start+length) % size;
+            int ps = (int) (start % size);
+            int es = (int) ((start+length) % size);
             if (ps < es)
             {
                 return new String(array, ps, length);
@@ -207,7 +207,7 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
         }
         if (array != null)
         {
-            int cms = cursor % size;
+            int cms = (int) (cursor % size);
             if (size - cms < text.length)
             {
                 System.arraycopy(text, 0, array, cms, size - cms);
@@ -262,8 +262,8 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
             int src;
             int dst;
             int len;
-            int ems = end % size;
-            int cms = cursor % size;
+            int ems = (int) (end % size);
+            int cms = (int) (cursor % size);
             if (ems < cms)
             {
                 src = 0;
@@ -298,7 +298,7 @@ public abstract class CharInput<I> extends Input<I, CharBuffer>
         }
         else
         {
-            int len = end-cursor;
+            long len = end-cursor;
             for (int ii=0;ii<len;ii++)
             {
                 set(end+ln-ii, get(end-ln));
